build: 좌표정밀도를 개선을 위한 해당 속성변경

+++ back/src/infrastructure/database/orm/sequelize/models/location.js
@@ -24,22 +24,22 @@ module.exports = class Location extends Sequelize.Model {
                 allowNull: false
             },
             min_x: {
-                type: Sequelize.FLOAT,
+                type: Sequelize.DOUBLE(),
                 allowNull: true
 
             },
             min_y: {
-                type: Sequelize.FLOAT,
+                type: Sequelize.DOUBLE(),
                 allowNull: true
 
             },
             max_x: {
-                type: Sequelize.FLOAT,
+                type: Sequelize.DOUBLE(),
                 allowNull: true
 
             },
             max_y: {
-                type: Sequelize.FLOAT,
+                type: Sequelize.DOUBLE(),
                 allowNull: true
 
             }
